Forth commit1

+++ app/src/main/java/com/example/androidjail/dwq.kt
@@ -1,4 +1,4 @@
 package com.example.androidjail
 
-class dwq(val a : Int, val b : Int) {
+class dwq(val a : Int, val b : Float) {
 }
